Refactor RateProviderRepository to use database and Flow.
- Introduce `RateProviderRepositoryImpl` to interact with `RateProviderDao` and `CurrencyRateDao`.
- Update `RateProviderRepository` interface methods to return `Flow` and add `suspend` modifier where necessary.
- Modify `RateProvider` ID type from `Int` to `Long`.
- Adapt `RateProviderConverter` to handle `Long` ID for `RateProviderEntity`.
- Update `KoinModules.kt` to provide `RateProviderRepositoryImpl` and DAOs.
- Add `deleteAllRates` to `CurrencyRateDao` and `deleteAllProviders` to `RateProviderDao`.
- Clear all database tables on app start in `RateManApp`.
- Temporarily disable API calls in `RatesViewModel` for `getRates` and `getRatesAndStore`.
- Make `RateProviderEntity` `name` unique.
- Initialize language in `MainActivity.onCreate` before `super.onCreate`.
- Comment out `RateProviderRepositoryFake`.

+++ app/src/main/java/com/currency/rateman/api/RateProviderConverter.kt
@@ -33,7 +33,7 @@ object RateProviderConverter {
 
     fun toRateProvider(entity: RateProviderEntity, rates: List<CurrencyRateEntity>): RateProvider {
         return RateProvider(
-            id = entity.id.toInt(),
+            id = entity.id.toLong(),
             name = entity.name,
             baseCurrency = CurrencyCode.valueOf(entity.baseCurrency),
             rates = rates.map { rate ->
